Default manifest for seL4 v11 iMX8+FPU verification.
This default.xml tracks the imx8-fpu-ver-v11 branches of both seL4 and
l4v for purposes of release to the external world. It will need to be
updated as part any support of this port.

The manifest targets a custom release of seL4 11.0.0 with backported
fixes and changes needed for verification to go through for iMX8 with
FPU enabled.

Signed-off-by: Rafal Kolanski <rafal.kolanski@data61.csiro.au>